[Silver V] Title: 타일 장식물, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13301. 타일 장식물/타일 장식물.cc
@@ -6,9 +6,9 @@ using namespace std;
 int main() {
 	cin.tie(0);
 	ios::sync_with_stdio(false);
-	int n;
+	long long int n;
 	cin >> n;
-	int* arr = new int[n + 1];
+	long long int* arr = new long long int[n + 1];
 	if (n == 0) {
 		cout << 0;
 		return 0;
